+++ src/routes/persona.js
@@ -41,7 +41,7 @@ router.post("/persona", async (req, res)=>{
     const {nombre, apellido, email, alias} = req.body;
     try{
         //valido que me envie correctamente la info
-        if(!nombre||!apellido||!email||!alias){
+        if (isEmpty(nombre) || isEmpty(apellido) || isEmpty(alias) || isEmpty(email)){
             throw new Error("Faltaron completar datos");
            }
        //verifico que no existe ese email previamente
@@ -56,10 +56,9 @@ router.post("/persona", async (req, res)=>{
        query = "INSERT INTO persona (nombre, apellido, alias, email) VALUE (?,?,?,?)";
        respuesta = await conexion.query(query, [nombre.toUpperCase(), apellido.toUpperCase(), alias.toUpperCase(), email]);
         
-       query = "SELECT * FROM persona WHERE nombre = ?";
-       respuesta = await conexion.query(query, [nombre])
+       query = "SELECT * FROM persona WHERE email = ?";
+       respuesta = await conexion.query(query, [email])
        res.status(200).send({"respuesta" : respuesta});
-
 }
 catch(e){
     console.error(e.message);
@@ -67,38 +66,42 @@ catch(e){
 }
 });
 
-// PUT '/persona/:id' recibe: {nombre: string, apellido: string, alias: string, email: string} el email no se puede modificar. retorna status 200 y el objeto modificado o bien status 413, {mensaje: <descripcion del error>} "error inesperado", "no se encuentra esa persona"
 
 router.put("/persona/:id" , async (req, res)=>{ //Para modificar una persona
     const {id} = req.params;
     const {nombre, apellido, email, alias} = req.body;
+
 try{
     if(email){
         throw new Error("El campo mail no se puede modificar")
     }
-    if(!nombre || !apellido || !alias){
-        throw new Error("No completaste los campos");
+    if(isEmpty(nombre) || isEmpty(apellido) || isEmpty(alias)){
+        throw new Error('No completaste los campos');
     }
 
     let query = "SELECT * FROM persona WHERE nombre = ? AND apellido = ? AND alias = ? AND id <> ?";
     let respuesta = await conexion.query(query, [nombre,apellido, alias, id]);
 
-   
     if(respuesta.length >0){
         throw new Error("La persona que queres ingresar ya existe");
     }
+    query = 'SELECT * FROM persona WHERE  id = ?';
+    respuesta = await conexion.query(query, [id]);
 
-    query = "UPDATE persona SET nombre = ?, apellido = ?, alias = ?";
-    respuesta = await conexion.query(query, [nombre.toUpperCase(),apellido.toUpperCase(), alias.toUpperCase()]);
+    if (respuesta.length == 0){
+        throw new Error("No se encuentra esa persona");
+    }
+
+    query = "UPDATE persona SET nombre = ?, apellido = ?, alias = ? WHERE id= ?";
+    respuesta = await conexion.query(query, [nombre.toUpperCase(),apellido.toUpperCase(), alias.toUpperCase(), id]);
 
     query = "SELECT * FROM persona WHERE id = ?";
     respuesta = await conexion.query(query, [id]);
+
     res.status(200).send({"Respuesta" : respuesta});
-   
    }
-   catch(error){
-       res.status(413).send({"Error" : "No se encuentra esa persona"});
-       
+catch(error){
+       res.status(413).send({"Error" : error.message});
    }
 });
 
@@ -107,10 +110,21 @@ try{
 router.delete("/persona/:id", async (req, res)=>{
     const {id} =req.params;
 try{
-     const query = "DELETE FROM persona WHERE id = ?";
+    let query = "SELECT * FROM persona WHERE id = ?";
+    let respuesta = await conexion.query(query, [id]);
+    
+    if (respuesta.length == 0) {
+       throw new Error("Esa persona no existe");
+    }
+    
+    if (respuesta.length > 0) {
+        throw new Error("Esta persona tiene libros asociados, no se puede borrar");
+    }
 
-     const respuesta = await conexion.query(query, [id]);
-     console.log(respuesta);
+    query = "DELETE FROM persona WHERE id = ?";
+
+     respuesta = await conexion.query(query, [id]);
+     
 
      res.status(200).send({"respuesta" : "Se borro correctamente"});
  }
@@ -119,5 +133,9 @@ try{
     res.status(413).send({"Error" : error.message});
 }
 });
+
+function isEmpty(str){
+    return (!str || 0 === str.trim().length);
+}
 
 module.exports = router;
